backend/requirements.txt
fastapi>=0.100.0
uvicorn>=0.20.0
httpx>=0.24.0
requests>=2.28.0
pydantic==1.10.13
python-multipart>=0.0.5
python-dotenv>=0.19.0 